Fixed maps

+++ ui/src/components/RequestList.tsx
@@ -66,7 +66,7 @@ const AdminRequestList: React.FC<Props> = ({requests, requestsId, update}) => {
                     >
                     </List.Item>
                     <div style={{width:"100%", height:"300px"}}>
-                        <PinsMap lat={parseInt(request.flight.x)} lng={parseInt(request.flight.y)}/>
+                        <PinsMap lat={parseFloat(request.flight.x)} lng={parseFloat(request.flight.y)}/>
                     </div>
                     <ViewContract
                         receivers={request.parties}
